Added divides and equals button functionality
The divide button dived two numbers. The process is that the first number is clicked, then the divides operator, then second number. After this process, if the divides is hit again, the result will be computed and displayed. The equals button carries out the operation the user wants. The process is after, say, the above divide process, equal is hit and the result pops up. If equal is hit again, the result will be the first number of a new operation with the same operator and second number.